Rename util files to fit patterns
